requirements.txt
fastapi
uvicorn
sqlalchemy
pandas
openai
python-jose[cryptography]
psycopg2-binary
chainlit
